move health route up to binary
