+++ rust/cloud-storage/contacts_service/Cargo.toml
@@ -17,16 +17,18 @@ path = "src/openapi.rs"
 anyhow = { workspace = true }
 aws-sdk-secretsmanager = { workspace = true }
 aws-sdk-sqs = { workspace = true }
+axum = { workspace = true }
 connection_gateway_client = { path = "../connection_gateway_client" }
 contacts = { path = "../contacts", features = ["axum"] }
 macro_auth = { path = "../macro_auth" }
 macro_aws_config = { path = "../macro_aws_config" }
+macro_cors = { path = "../macro_cors" }
 macro_entrypoint = { path = "../macro_entrypoint" }
-model = { path = "../model" }
 macro_env = { path = "../macro_env" }
 macro_middleware = { path = "../macro_middleware", default-features = false, features = [
   "auth",
 ] }
+model = { path = "../model" }
 rate_limit = { path = "../rate_limit", features = ["axum"] }
 redis = { workspace = true }
 secretsmanager_client = { path = "../secretsmanager_client" }
@@ -37,6 +39,11 @@ sqs_worker = { path = "../sqs_worker" }
 tokio = { workspace = true }
 tracing = { workspace = true }
 utoipa = { workspace = true }
+utoipa-swagger-ui = { workspace = true }
 
 [features]
 mock = []
+
+[dev-dependencies]
+http-body-util = { workspace = true }
+tower = { workspace = true }
