+++ rust/cloud-storage/contacts_service/src/health.rs
